feat!: counts/sizes return number, not bigint
Drop accidental JS bigints from count/size getters so ordinary arithmetic
works without a "Cannot mix BigInt" TypeError.

- Blob.size(): u64 -> i64 (byte size, lossless to 2^53)
- DiffFile.size(): u64 -> i64
- Commit.parentCount(): usize -> u32
- Tree.len() -> Tree.size(): u64 -> u32 (rename also avoids
  clippy::len_without_is_empty)

napi mapping: i64/u32 -> number; u64/usize -> bigint. index.d.ts regenerated
from the napi build; index.js loader is unaffected. Adds counts-sizes.spec.mjs
asserting typeof number + arithmetic for all four getters.

+++ src/commit.rs
@@ -244,8 +244,8 @@ impl Commit {
   /// Get the number of parents of this commit.
   ///
   /// Use the `parents` iterator to return an iterator over all parents.
-  pub fn parent_count(&self) -> usize {
-    self.inner.parent_count()
+  pub fn parent_count(&self) -> u32 {
+    self.inner.parent_count() as u32
   }
 
   #[napi]
